-[REFACTOR] Upgrade, move and rename the DataSource interface.
