+++ build.gradle
@@ -16,11 +16,11 @@ allprojects {
 }
 
 // - P R O J E C T   D E F I N I T I O N
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(17)
-    }
-}
+//java {
+//    toolchain {
+//        languageVersion = JavaLanguageVersion.of(17)
+//    }
+//}
 group = 'org.dimensinfin.neocom'
 version = '0.29.4'
 archivesBaseName = 'neocom-datamanagement'
