+++ dependencies.gradle
@@ -14,11 +14,11 @@ dependencies {
     annotationProcessor "org.projectlombok:lombok:$lombokVersion"
     implementation "org.projectlombok:lombok:$lombokVersion"
     implementation "org.slf4j:slf4j-api:$slf4jVersion"
-    implementation "org.jetbrains:annotations:20.1.0"
+    implementation "org.jetbrains:annotations:24.1.0"
     implementation "joda-time:joda-time:$jodaTimeVersion" // Use this library for support on Android systems
     implementation 'org.apache.commons:commons-lang3:3.14.0'
     implementation 'com.annimon:stream:1.2.2' // Android compatible streams
-    implementation 'com.google.inject:guice:4.2.3' // Google injection library compatible with Android
+    implementation 'com.google.inject:guice:5.1.0' // Google injection library compatible with Android
 //    implementation 'org.redisson:redisson:3.15.0' // Redisson Redis client to implements second level cache
 }
 
@@ -39,7 +39,7 @@ dependencies {
     implementation "com.nytimes.android:cache3:$cacheVersion"
     implementation "com.nytimes.android:middleware3:$cacheVersion"
     implementation "com.nytimes.android:filesystem3:$cacheVersion"
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.12'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.21'
     implementation 'com.jakewharton:disklrucache:2.0.2'
     implementation 'com.jcabi:jcabi-aspects:0.22.6' // Useful annotations for methods
     implementation 'org.modelmapper:modelmapper:3.2.0' // Help mapping ESI Get200Ok classes to similar classes on the app model
