+++ src/integration/java/org/dimensinfin/eveonline/neocom/IntegrationNeoComServicesDependenciesModule.java
@@ -14,7 +14,7 @@ import org.dimensinfin.eveonline.neocom.provider.IConfigurationService;
 import org.dimensinfin.eveonline.neocom.provider.IFileSystem;
 import org.dimensinfin.eveonline.neocom.service.DMServicesDependenciesModule;
 import org.dimensinfin.eveonline.neocom.service.ESIDataService;
-import org.dimensinfin.eveonline.neocom.service.IDataStore;
+import org.dimensinfin.eveonline.neocom.ports.IDataStorePort;
 import org.dimensinfin.eveonline.neocom.service.IStoreCache;
 import org.dimensinfin.eveonline.neocom.service.LocationCatalogService;
 import org.dimensinfin.eveonline.neocom.service.MemoryStoreCacheService;
@@ -109,7 +109,7 @@ public class IntegrationNeoComServicesDependenciesModule extends AbstractModule 
 				.to( MarketService.class )
 				.in( Singleton.class );
 
-		this.bind( IDataStore.class )
+		this.bind( IDataStorePort.class )
 				.annotatedWith( Names.named( DMServicesDependenciesModule.IDATA_STORE ) )
 				.to( IntegrationRedisDataStoreImplementation.class )
 				.in( Singleton.class );
