+++ src/integration/java/org/dimensinfin/eveonline/neocom/service/LocationCatalogServiceIT.java
@@ -9,6 +9,7 @@ import org.testcontainers.containers.GenericContainer;
 
 import org.dimensinfin.eveonline.neocom.IntegrationNeoComServicesDependenciesModule;
 import org.dimensinfin.eveonline.neocom.esiswagger.model.GetAlliancesAllianceIdOk;
+import org.dimensinfin.eveonline.neocom.ports.IDataStorePort;
 import org.dimensinfin.eveonline.neocom.ports.ILocationFactoryPort;
 import org.dimensinfin.eveonline.neocom.provider.IFileSystem;
 import org.dimensinfin.eveonline.neocom.support.IntegrationRedisDataStoreImplementation;
@@ -33,7 +34,7 @@ public class LocationCatalogServiceIT {
 	private IStoreCache storeCache;
 	private RetrofitService retrofitService;
 	private LocationCatalogService locationCatalogService;
-	private IDataStore dataStore;
+	private IDataStorePort dataStore;
 	private ESIDataService esiDataService;
 	private ILocationFactoryPort locationFactory;
